+++ clash-lib/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for the [`clash-lib`](http://hackage.haskell.org/package/clash-lib) package
 
+## 0.6.1
+* New features:
+  * Support for `clash-prelude` 0.10.1
+
 ## 0.6 *October 3rd 2015*
 * New features:
   * Support for `clash-prelude` 0.10