+++ clash-verilog/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for the [`clash-systemverilog`](http://hackage.haskell.org/package/clash-systemverilog) package
 
+## 0.6.1
+* New features:
+  * Support for `clash-prelude` 0.10.1
+
 ## 0.6
 * New features:
   * Support `clash-prelude-0.10`